backend/requirements.txt
anthropic>=0.40.0
python-frontmatter>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
